# encoding: Return to base-100 encoding for Decimal