--- a/roachpb/data.go
+++ b/roachpb/data.go
@@ -446,9 +446,9 @@ func (v *Value) SetDuration(t duration.Duration) error {
 // SetDecimal encodes the specified decimal value into the bytes field of
 // the receiver using Gob encoding, sets the tag and clears the checksum.
 func (v *Value) SetDecimal(dec *inf.Dec) error {
-	encodingSizeOverestimate := headerSize + encoding.OverestimateDecimalSize(dec)
-	v.RawBytes = make([]byte, encodingSizeOverestimate)
-	v.RawBytes = encoding.EncodeDecimalAscending(v.RawBytes[:headerSize], dec)
+	decSize := encoding.UpperBoundNonsortingDecimalSize(dec)
+	v.RawBytes = make([]byte, headerSize, headerSize+decSize)
+	v.RawBytes = encoding.EncodeNonsortingDecimal(v.RawBytes, dec)
 	v.setTag(ValueType_DECIMAL)
 	return nil
 }
@@ -536,8 +536,7 @@ func (v Value) GetDecimal() (*inf.Dec, error) {
 	if tag := v.GetTag(); tag != ValueType_DECIMAL {
 		return nil, fmt.Errorf("value type is not %s: %s", ValueType_DECIMAL, tag)
 	}
-	_, d, err := encoding.DecodeDecimalAscending(v.dataBytes(), nil)
-	return d, err
+	return encoding.DecodeNonsortingDecimal(v.dataBytes(), nil)
 }
 
 // GetTimeseries decodes an InternalTimeSeriesData value from the bytes