--- a/util/encoding/decimal_test.go
+++ b/util/encoding/decimal_test.go
@@ -20,6 +20,7 @@ import (
 	"bytes"
 	"fmt"
 	"math"
+	"math/big"
 	"math/rand"
 	"strconv"
 	"strings"
@@ -216,7 +217,7 @@ func TestEncodeDecimalRand(t *testing.T) {
 		for _, tmp := range [][]byte{nil, make([]byte, 0, 100)} {
 			// Test with and without appending buffer.
 			for _, append := range [][]byte{nil, []byte("hello")} {
-				const randomTrials = 1000
+				const randomTrials = 50000
 				for i := 0; i < randomTrials; i++ {
 					cur := decimal.NewDecFromFloat(rand.Float64())
 
@@ -241,11 +242,6 @@ func TestEncodeDecimalRand(t *testing.T) {
 						t.Fatalf("unexpected mismatch for %v, got %v", cur, res)
 					}
 
-					// Make sure we would have overestimated the value.
-					if est := OverestimateDecimalSize(cur); est < len(enc) {
-						t.Fatalf("expected estimate of %d for %v to be greater than or equal to the encoded length, found [% x]", est, cur, enc)
-					}
-
 					// Make sure lexicographical sorting is consistent.
 					if prev != nil {
 						bytesCmp := bytes.Compare(prevEnc, enc)
@@ -263,6 +259,155 @@ func TestEncodeDecimalRand(t *testing.T) {
 					prevEnc = enc
 				}
 			}
+		}
+	}
+}
+
+func TestNonsortingEncodeDecimal(t *testing.T) {
+	testCases := []struct {
+		Value    *inf.Dec
+		Encoding []byte
+	}{
+		{inf.NewDec(-99122, -99999), []byte{0x22, 0xf8, 0x01, 0x86, 0xa4, 0x01, 0x83, 0x32}},
+		// Three duplicates to make sure -13*10^1000 <= -130*10^999 <= -13*10^1000
+		{inf.NewDec(-13, -1000), []byte{0x22, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(-130, -999), []byte{0x22, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(-13, -1000), []byte{0x22, 0xf7, 0x03, 0xea, 0x0d}},
+		{decimal.NewDecFromFloat(-math.MaxFloat64), []byte{0x22, 0xf7, 0x01, 0x35, 0x3f, 0xdd, 0xec, 0x7f, 0x2f, 0xaf, 0x35}},
+		{inf.NewDec(-130, -100), []byte{0x22, 0xef, 0x0d}},
+		{inf.NewDec(-13, 0), []byte{0x22, 0x8a, 0x0d}},
+		{inf.NewDec(-11, 0), []byte{0x22, 0x8a, 0x0b}},
+		{inf.NewDec(-1, 0), []byte{0x22, 0x89, 0x01}},
+		{inf.NewDec(-8, 1), []byte{0x2d, 0x08}},
+		{inf.NewDec(-1, 1), []byte{0x2d, 0x01}},
+		{inf.NewDec(-11, 4), []byte{0x2e, 0x8a, 0x0b}},
+		{inf.NewDec(-11, 6), []byte{0x2e, 0x8c, 0x0b}},
+		{decimal.NewDecFromFloat(-math.SmallestNonzeroFloat64), []byte{0x2e, 0xf7, 0x01, 0x43, 0x05}},
+		{inf.NewDec(-11, 66666), []byte{0x2e, 0xf8, 0x01, 0x04, 0x68, 0x0b}},
+		{inf.NewDec(0, 0), []byte{0x2f}},
+		{decimal.NewDecFromFloat(math.SmallestNonzeroFloat64), []byte{0x30, 0xf7, 0x01, 0x43, 0x05}},
+		{inf.NewDec(11, 6), []byte{0x30, 0x8c, 0x0b}},
+		{inf.NewDec(11, 4), []byte{0x30, 0x8a, 0x0b}},
+		{inf.NewDec(1, 1), []byte{0x31, 0x01}},
+		{inf.NewDec(12345, 5), []byte{0x31, 0x30, 0x39}},
+		{inf.NewDec(8, 1), []byte{0x31, 0x08}},
+		{inf.NewDec(1, 0), []byte{0x3c, 0x89, 0x01}},
+		{inf.NewDec(11, 0), []byte{0x3c, 0x8a, 0x0b}},
+		{inf.NewDec(13, 0), []byte{0x3c, 0x8a, 0x0d}},
+		// Note that this does not sort correctly!
+		{inf.NewDec(255, 0), []byte{0x3c, 0x8b, 0xff}},
+		{inf.NewDec(256, 0), []byte{0x3c, 0x8b, 0x01, 0x00}},
+		{decimal.NewDecFromFloat(math.MaxFloat64), []byte{0x3c, 0xf7, 0x01, 0x35, 0x3f, 0xdd, 0xec, 0x7f, 0x2f, 0xaf, 0x35}},
+		// Four duplicates to make sure 13*10^1000 <= 130*10^999 <= 1300*10^998 <= 13*10^1000
+		{inf.NewDec(13, -1000), []byte{0x3c, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(130, -999), []byte{0x3c, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(1300, -998), []byte{0x3c, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(13, -1000), []byte{0x3c, 0xf7, 0x03, 0xea, 0x0d}},
+		{inf.NewDec(99122, -99999), []byte{0x3c, 0xf8, 0x01, 0x86, 0xa4, 0x01, 0x83, 0x32}},
+		{inf.NewDec(99122839898321208, -99999), []byte{0x3c, 0xf8, 0x01, 0x86, 0xb0, 0x01, 0x60, 0x27, 0xb2, 0x9d, 0x44, 0x71, 0x38}},
+	}
+
+	for _, tmp := range [][]byte{nil, make([]byte, 0, 100)} {
+		for i, c := range testCases {
+			enc := EncodeNonsortingDecimal(nil, c.Value)
+			dec, err := DecodeNonsortingDecimal(enc, tmp)
+			if err != nil {
+				t.Error(err)
+				continue
+			}
+			if !bytes.Equal(enc, c.Encoding) {
+				t.Errorf("unexpected mismatch for %s. expected [% x], got [% x]",
+					c.Value, c.Encoding, enc)
+			}
+			if dec.Cmp(c.Value) != 0 {
+				t.Errorf("%d unexpected mismatch for %v. got %v", i, c.Value, dec)
+			}
+		}
+
+		// Test that appending the decimal to an existing buffer works.
+		other := inf.NewDec(123, 2)
+		enc := EncodeNonsortingDecimal([]byte("hello"), other)
+		dec, err := DecodeNonsortingDecimal(enc[5:], tmp)
+		if err != nil {
+			t.Error(err)
+			continue
+		}
+		if dec.Cmp(other) != 0 {
+			t.Errorf("unexpected mismatch for %v. got %v", 1.23, other)
+		}
+	}
+}
+
+func TestNonsortingEncodeDecimalRand(t *testing.T) {
+	// Test with and without tmp buffer.
+	for _, tmp := range [][]byte{nil, make([]byte, 0, 100)} {
+		// Test with and without appending buffer.
+		for _, append := range [][]byte{nil, []byte("hello")} {
+			const randomTrials = 200000
+			for i := 0; i < randomTrials; i++ {
+				cur := decimal.NewDecFromFloat(rand.Float64())
+
+				enc := EncodeNonsortingDecimal(append, cur)
+				enc = enc[len(append):]
+				res, err := DecodeNonsortingDecimal(enc, tmp)
+				if err != nil {
+					t.Fatal(err)
+				}
+
+				// Make sure we decode the same value we encoded.
+				if cur.Cmp(res) != 0 {
+					t.Fatalf("unexpected mismatch for %v, got %v", cur, res)
+				}
+
+				// Make sure we would have overestimated the value.
+				if est := UpperBoundNonsortingDecimalSize(cur); est < len(enc) {
+					t.Fatalf("expected estimate of %d for %v to be greater than or equal to the encoded length, found [% x]", est, cur, enc)
+				}
+			}
+		}
+	}
+}
+
+func TestDigitsLookupTable(t *testing.T) {
+	// Make sure all elements in table make sense.
+	min := new(big.Int)
+	prevBorder := big.NewInt(0)
+	for i := 1; i <= tableSize; i++ {
+		elem := digitsLookupTable[i]
+
+		min.SetInt64(2)
+		min.Exp(min, big.NewInt(int64(i-1)), nil)
+		if minLen := len(min.String()); minLen != elem.digits {
+			t.Errorf("expected 2^%d to have %d digits, found %d", i, elem.digits, minLen)
+		}
+
+		if zeros := strings.Count(elem.border.String(), "0"); zeros != elem.digits {
+			t.Errorf("the %d digits for digitsLookupTable[%d] does not agree with the border %v", elem.digits, i, &elem.border)
+		}
+
+		if min.Cmp(&elem.border) >= 0 {
+			t.Errorf("expected 2^%d = %v to be less than the border, found %v", i-1, min, &elem.border)
+		}
+
+		if elem.border.Cmp(prevBorder) > 0 {
+			if min.Cmp(prevBorder) <= 0 {
+				t.Errorf("expected 2^%d = %v to be greater than or equal to the border, found %v", i-1, min, prevBorder)
+			}
+			prevBorder = &elem.border
+		}
+	}
+
+	// Throw random big.Ints at the table and make sure the
+	// digit lengths line up.
+	const randomTrials = 100
+	for i := 0; i < randomTrials; i++ {
+		a := big.NewInt(rand.Int63())
+		b := big.NewInt(rand.Int63())
+		a.Mul(a, b)
+
+		tableDigits, _ := numDigits(a, nil)
+		if actualDigits := len(a.String()); actualDigits != tableDigits {
+			t.Errorf("expected %d digits for %v, found %d", tableDigits, a, actualDigits)
 		}
 	}
 }
@@ -297,5 +442,38 @@ func BenchmarkDecodeDecimal(b *testing.B) {
 	b.ResetTimer()
 	for i := 0; i < b.N; i++ {
 		_, _, _ = DecodeDecimalAscending(vals[i%len(vals)], buf)
+	}
+}
+
+func BenchmarkNonsortingEncodeDecimal(b *testing.B) {
+	rng, _ := randutil.NewPseudoRand()
+
+	vals := make([]*inf.Dec, 10000)
+	for i := range vals {
+		vals[i] = decimal.NewDecFromFloat(rng.Float64())
+	}
+
+	buf := make([]byte, 0, 100)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_ = EncodeNonsortingDecimal(buf, vals[i%len(vals)])
+	}
+}
+
+func BenchmarkNonsortingDecodeDecimal(b *testing.B) {
+	rng, _ := randutil.NewPseudoRand()
+
+	vals := make([][]byte, 10000)
+	for i := range vals {
+		d := decimal.NewDecFromFloat(rng.Float64())
+		vals[i] = EncodeNonsortingDecimal(nil, d)
+	}
+
+	buf := make([]byte, 0, 100)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		_, _ = DecodeNonsortingDecimal(vals[i%len(vals)], buf)
 	}
 }