cc. @bdarnell 

Closes #5965.

Again unfortunately this is a breaking change. I discussed with Ben whether
we should provide a migration path

The previous encoding approach for DECIMAL was broken in two ways:
- It prefixed the mantissa with the mantissa's length, which meant that
  decimal values did not always sort correctly. For instance `.01111`
  would sort before `0.02`.
- It encoded its exponent in a base that was not a factor of its
  mantissa's base, which also permitted very subtle ordering errors.
  Here, the encoding's exponent was base-10 encoded while
  its mantissa was base-256 encoded. This was not a problem in the
  current iteration of the previous encoding because the mantissa length
  varint (discussed above) would prevent any strange ordering. However, if
  we had moved from the mantissa length prefix approach to the escaped
  0x00 byte approach for encoding the mantissa, this issue would have
  come up. In practice, it means that values like `255 = [0xff]` would
  sort before `256 = [0x01 0x00]`. It is critical that the exponent and
  the mantissa share the same base (or the exponent is a factor of the
  mantissa's base) so that the exponent will increment whenever the
  mantissa's byte size grows.

I spent all day trying to convert the base-10 exponent into a base-256 (or
base-16) exponent so we could avoid marshaling through strings on encoding
and decoding, but had little luck. If anyone has ideas, I'd love to hear
them.

The current approach uses a base-100 exponent and a base-100 mantissa.
It is similar to the first encoding of decimal, with a few key
differences.
- The null terminator is searched for AFTER the uvarint has been
  decoded. This was an issue with both the previous float and decimal
  encodings, which both assumed that the exponent uvarint would be free
  of 0x00 bytes. In those instances, there was a risk of searching for
  the terminator, only to find a 0x00 byte in the uvarint.
- A few performance tweaks were made to reduce an allocation from
  encoding and to reduce an allocation from decoding.
- We're now using the `{Encode/Decode}Uvarint{Ascending/Descending}`
  functions instead of the old uvarint functions.
- I've added a random encoding test to make sure this is thoroughly correct. 
  So far, so good...

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="35" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/cockroachdb/cockroach/5994)
<!-- Reviewable:end -->