@knz 2016-04-12
Woah using base-100 is a huge step backwards in terms of efficiency. 
Each digit encodes in 6.64 bits which is 83% storage efficiency per byte.

What about base 160, which encodes in bytes at 91%, or base 40960 which encodes in digits of 16 bits at 96% efficiency?