@nvanbenschoten 2016-04-12
The result of this change on performance is along the lines of:

```
name               old time/op    new time/op    delta
ValueSetDecimal-4     199ns ± 2%     181ns ± 1%    -9.17%        (p=0.000 n=20+10)
ValueGetDecimal-4     218ns ± 2%     196ns ± 1%   -10.07%        (p=0.000 n=20+10)
EncodeDecimal-4       155ns ± 3%     396ns ± 2%  +156.47%        (p=0.000 n=10+10)
DecodeDecimal-4       206ns ± 2%     504ns ± 1%  +143.96%         (p=0.000 n=10+9)

name               old alloc/op   new alloc/op   delta
ValueSetDecimal-4     32.0B ± 0%     32.0B ± 0%      ~     (all samples are equal)
ValueGetDecimal-4     96.0B ± 0%     96.0B ± 0%      ~     (all samples are equal)
EncodeDecimal-4      0.00B ±NaN%    69.00B ± 0%     +Inf%        (p=0.000 n=10+10)
DecodeDecimal-4       96.0B ± 0%    128.0B ± 0%   +33.33%        (p=0.000 n=10+10)

name               old allocs/op  new allocs/op  delta
ValueSetDecimal-4      1.00 ± 0%      1.00 ± 0%      ~     (all samples are equal)
ValueGetDecimal-4      2.00 ± 0%      2.00 ± 0%      ~     (all samples are equal)
EncodeDecimal-4       0.00 ±NaN%      2.00 ± 0%     +Inf%        (p=0.000 n=10+10)
DecodeDecimal-4        2.00 ± 0%      3.00 ± 0%   +50.00%        (p=0.000 n=10+10)
```

Note that decimal key encoding got slower but decimal value encoding actually got faster (and smaller).