@petermattis 2016-04-13
Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions.

---

*[util/encoding/decimal_test.go, line 220 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEtmC5nBXhQZiErSAc:-KFEtmC5nBXhQZiErSAd:-1410151456) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/decimal_test.go#L220)):*
How long does this test take? I would bump this up until the test takes ~1s.

---

*[util/encoding/encoding.go, line 48 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEu9ayzZwHMNIyvG2L:-KFEu9ayzZwHMNIyvG2M:1724353026) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/encoding.go#L48)):*
Uh-oh! We don't want to change `bytesMarker` and all of the other markers. Doing so will break compatibility for all encoded data.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994)*
<!-- Sent from Reviewable.io -->