@petermattis 2016-04-13
Let's add another test which encodes a value for each of the 

---

Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions.

---

*[util/encoding/encoding.go, line 48 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEu9ayzZwHMNIyvG2L:-KFEuwwtogQvbZbSK__q:552054498) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/encoding.go#L48)):*
Err, never mind. I see that you're just hardcoding the value instead of deriving it from the `decimal` markers. Carry on.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-:-KFEuXBOpUMQvZADzixG:-1121374696)*
<!-- Sent from Reviewable.io -->