@petermattis 2016-04-13
My comment was cut off. I was going to suggest that we add a test that uses each of the encoding routines, hashes (sha1) the output and verifies against golden data, but we more or less have that with the various unit tests.

---

Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-:-KFEv3xTDE_E4otKwd2Z:47817054)*
<!-- Sent from Reviewable.io -->