@petermattis 2016-04-13
<img class="emoji" title=":lgtm:" alt=":lgtm:" align="absmiddle" src="https://reviewable.io/lgtm.png" height="20" width="61"/>

---

Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions.

---



*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-:-KFEvOQr4s8cZhJCBm0P:1417088721)*
<!-- Sent from Reviewable.io -->