@bdarnell 2016-04-13
Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions, all commit checks successful.

---

*[util/encoding/encoding.go, line 48 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEu9ayzZwHMNIyvG2L:-KFFJCWL_ohCJrn9XDD1:1964836411) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/encoding.go#L48)):*
I think it's fine to move towards hard-coding constants here (we don't want to make these values easy to change accidentally), but there should be a comment about this gap in the sequence (and the gap between `durationBigPosMarker` and `decimalNan`. why not start `decimalNan` at `durationBigPosMarker+1`?)

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994)*
<!-- Sent from Reviewable.io -->