@nvanbenschoten 2016-04-13
Review status: 0 of 5 files reviewed at latest revision, 2 unresolved discussions.

---

*[util/encoding/decimal_test.go, line 220 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEtmC5nBXhQZiErSAc:-KFFvtXrZgFzkB7NG3Ap:1084272768) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/decimal_test.go#L220)):*
I bumped both of the random tests to take about a second.

---

*[util/encoding/encoding.go, line 48 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEu9ayzZwHMNIyvG2L:-KFFw6HEr6x_hMFN4dyo:512491286) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/encoding.go#L48)):*
Added a comment.

And I was hoping to leave some room for "growth". The reason I had to move the markers around here was because there was no room to add an extra tag. I dont feel super strongly about this though. Do you think I should push the decimal markers up against `durationBigPosMarker`?

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994)*
<!-- Sent from Reviewable.io -->