@bdarnell 2016-04-13
Review status: 0 of 5 files reviewed at latest revision, 1 unresolved discussion.

---

*[util/encoding/encoding.go, line 48 \[r2\]](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994#-KFEu9ayzZwHMNIyvG2L:-KFG2ZJkIK6PnwSpldml:1773495247) ([raw file](https://github.com/cockroachdb/cockroach/blob/880bf90c4958c2daa8fe293e80bf97f2190f8ab3/util/encoding/encoding.go#L48)):*
I'm thinking about fragmentation: this change relied on the fact that there was a large gap between `durationBigPosMarker` and `IntMin`. You've left a small gap which gives duration or decimal a little room to expand, but at the expense of space in the larger unclaimed area which might be needed for future data types that need a chunk of contiguous space. I don't feel strongly about it but I'd prefer to pack the types closely. 

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/5994)*
<!-- Sent from Reviewable.io -->